Fix bugs in Channel so that #close can be called twice
The fix should be much convenient comparing to Go's channel

Remove redundant iota

Update error messages in Channel class

Move tests to channel_test.go

Update vm/error.go
